Split PlayerInfo to PlayerInfo, PlayerInfoConstants, VictoryTypeByPointsConstants.
