+++ sources/Colonization/AngelScriptBinders/Core/PlayerInfo/BindPlayerInfo.cpp
@@ -2,54 +2,15 @@
 #include "BindPlayerInfo.hpp"
 #include <Urho3D/ThirdParty/AngelScript/angelscript.h>
 #include <Urho3D/AngelScript/APITemplates.h>
-#include <Colonization/Core/PlayerInfo.hpp>
+#include <Colonization/Core/PlayerInfo/PlayerInfo.hpp>
 
 namespace Colonization
 {
-// To register constants to AngelScript, we have to create getters for them.
-Urho3D::StringHash GetConst_PLAYER_INFO_VICTORY_TYPE_NAME_KEY ()
-{
-    return PLAYER_INFO_VICTORY_TYPE_NAME_KEY;
-}
-
-Urho3D::StringHash GetConst_PLAYER_INFO_VICTORY_TYPE_INFO_KEY ()
-{
-    return PLAYER_INFO_VICTORY_TYPE_INFO_KEY;
-}
-
-
-Urho3D::StringHash GetConst_PLAYER_INFO_VICTORY_TYPE_PROGRESS_KEY ()
-{
-    return PLAYER_INFO_VICTORY_TYPE_PROGRESS_KEY;
-}
-
-Urho3D::StringHash GetConst_VICTORY_TYPE_BY_POINTS ()
-{
-    return VICTORY_TYPE_BY_POINTS;
-}
-
-Urho3D::String GetConst_VICTORY_TYPE_BY_POINTS_NAME ()
-{
-    return VICTORY_TYPE_BY_POINTS_NAME;
-}
-
-Urho3D::String GetConst_VICTORY_TYPE_BY_POINTS_INFO ()
-{
-    return VICTORY_TYPE_BY_POINTS_INFO;
-}
-
 void BindPlayerInfo (Urho3D::Script *script)
 {
     asIScriptEngine *engine = script->GetScriptEngine ();
     Urho3D::RegisterComponent <PlayerInfo> (engine, "PlayerInfo");
     BindPlayerInfoInterface (script, "PlayerInfo");
-
-    engine->RegisterGlobalFunction ("StringHash get_PLAYER_INFO_VICTORY_TYPE_NAME_KEY ()", asFUNCTION (GetConst_PLAYER_INFO_VICTORY_TYPE_NAME_KEY), asCALL_CDECL);
-    engine->RegisterGlobalFunction ("StringHash get_PLAYER_INFO_VICTORY_TYPE_INFO_KEY ()", asFUNCTION (GetConst_PLAYER_INFO_VICTORY_TYPE_INFO_KEY), asCALL_CDECL);
-    engine->RegisterGlobalFunction ("StringHash get_PLAYER_INFO_VICTORY_TYPE_PROGRESS_KEY ()", asFUNCTION (GetConst_PLAYER_INFO_VICTORY_TYPE_PROGRESS_KEY), asCALL_CDECL);
-    engine->RegisterGlobalFunction ("StringHash get_VICTORY_TYPE_BY_POINTS ()", asFUNCTION (GetConst_VICTORY_TYPE_BY_POINTS), asCALL_CDECL);
-    engine->RegisterGlobalFunction ("String get_VICTORY_TYPE_BY_POINTS_NAME ()", asFUNCTION (GetConst_VICTORY_TYPE_BY_POINTS_NAME), asCALL_CDECL);
-    engine->RegisterGlobalFunction ("String get_VICTORY_TYPE_BY_POINTS_INFO ()", asFUNCTION (GetConst_VICTORY_TYPE_BY_POINTS_INFO), asCALL_CDECL);
 }
 
 void BindPlayerInfoInterface (Urho3D::Script *script, Urho3D::String className)
