+++ sources/Colonization/AngelScriptBinders/Core/PlayerInfo/BindPlayerInfo.hpp
