+++ sources/Colonization/AngelScriptBinders/Core/PlayerInfo/BindPlayerInfoConstants.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindPlayerInfoConstants (Urho3D::Script *script);
+}
