+++ sources/Colonization/AngelScriptBinders/Core/PlayerInfo/BindVictoryTypeByPointsConstants.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindVictoryTypeByPointsConstants (Urho3D::Script *script);
+}
