+++ sources/Colonization/Core/PlayerInfo/PlayerInfoConstants.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include <Urho3D/Scene/Component.h>
+
+namespace Colonization
+{
+const Urho3D::StringHash PLAYER_INFO_VICTORY_TYPE_NAME_KEY ("Name");
+const Urho3D::StringHash PLAYER_INFO_VICTORY_TYPE_INFO_KEY ("Info");
+const Urho3D::StringHash PLAYER_INFO_VICTORY_TYPE_PROGRESS_KEY ("Progress");
+}
