+++ sources/Colonization/Core/PlayerInfo/VictoryTypeByPointsConstants.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include <Urho3D/Scene/Component.h>
+
+namespace Colonization
+{
+const Urho3D::StringHash VICTORY_TYPE_BY_POINTS ("ByPoints");
+const Urho3D::String VICTORY_TYPE_BY_POINTS_NAME ("by points");
+const Urho3D::String VICTORY_TYPE_BY_POINTS_INFO ("Maximum game time elapsed. ${PlayerName} has biggest number of points.");
+}
